Add fgwas inital pruning step to run.qmd. Hacky meta-sumstats hardcoding.
Improved speed of fgwas model selection greatly by pruning any annotations which aren't at least nominally significant when in a model alone. This before, every permutation of every annotation would be tested. Added a bunch of temporary hard-coding to switch to using the meta-analyzed sumstats in fgwas, rather than running it separately for each ancestry; it can be assumed that the effect of annotations is consistent across ancestries.